requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
flask>=3.0.0
python-dotenv>=1.0.0
